Always use unsigned long for size_t

+++ source/rk_types.h
@@ -80,7 +80,7 @@ typedef uint8_t u8;
 typedef uint16_t u16;
 typedef uint32_t u32;
 typedef uint64_t u64;
-typedef uint32_t size_t;
+typedef unsigned long size_t;
 
 typedef volatile u8 vu8;
 typedef volatile u16 vu16;
